requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiofiles==23.2.1
    # via
    #   -r requirements.in
    #   alibabacloud-credentials
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.16
    # via alibabacloud-tea
aiosignal==1.3.2
    # via aiohttp
alembic==1.13.3
    # via -r requirements.in
alibabacloud-credentials==1.0.0
    # via
    #   alibabacloud-gateway-spi
    #   alibabacloud-tea-openapi
alibabacloud-credentials-api==1.0.0
    # via alibabacloud-credentials
alibabacloud-dysmsapi20170525==3.1.2
    # via -r requirements.in
alibabacloud-endpoint-util==0.0.3
    # via alibabacloud-dysmsapi20170525
alibabacloud-gateway-spi==0.0.3
    # via alibabacloud-tea-openapi
alibabacloud-openapi-util==0.2.2
    # via
    #   alibabacloud-dysmsapi20170525
    #   alibabacloud-tea-openapi
alibabacloud-tea==0.4.3
    # via
    #   alibabacloud-credentials
    #   alibabacloud-endpoint-util
    #   alibabacloud-tea-util
    #   alibabacloud-tea-xml
alibabacloud-tea-openapi==0.3.13
    # via
    #   -r requirements.in
    #   alibabacloud-dysmsapi20170525
alibabacloud-tea-util==0.3.13
    # via
    #   alibabacloud-dysmsapi20170525
    #   alibabacloud-openapi-util
    #   alibabacloud-tea-openapi
alibabacloud-tea-xml==0.0.2
    # via alibabacloud-tea-openapi
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
apscheduler==3.11.0
    # via alibabacloud-credentials
argon2-cffi==23.1.0
    # via -r requirements.in
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asyncpg==0.29.0
    # via -r requirements.in
attrs==25.3.0
    # via aiohttp
babel==2.13.1
    # via -r requirements.in
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via
    #   argon2-cffi-bindings
    #   cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
    #   uvicorn
cryptography==44.0.2
    # via
    #   alibabacloud-openapi-util
    #   python-jose
deprecated==1.2.18
    # via limits
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
email-validator==2.2.0
    # via -r requirements.in
fastapi==0.115.12
    # via -r requirements.in
fastapi-cli==0.0.7
    # via -r requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.8
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   -r requirements.in
    #   openai
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
itsdangerous==2.2.0
    # via -r requirements.in
jiter==0.9.0
    # via openai
limits==4.7.3
    # via slowapi
mako==1.3.10
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via mako
mdurl==0.1.2
    # via markdown-it-py
multidict==6.4.3
    # via
    #   aiohttp
    #   yarl
numpy==2.2.4
    # via
    #   -r requirements.in
    #   pandas
openai==1.74.0
    # via -r requirements.in
orjson==3.10.16
    # via -r requirements.in
packaging==24.2
    # via limits
pandas==2.2.3
    # via -r requirements.in
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
psycopg2==2.9.10
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic==2.10.6
    # via
    #   -r requirements.in
    #   fastapi
    #   openai
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.2.1
    # via -r requirements.in
pygments==2.19.1
    # via rich
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   pandas
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-jose[cryptography]==3.4.0
    # via -r requirements.in
python-json-logger==2.0.7
    # via -r requirements.in
python-magic==0.4.27
    # via -r requirements.in
python-multipart==0.0.20
    # via -r requirements.in
python-ulid==2.2.0
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via uvicorn
redis==5.2.1
    # via -r requirements.in
requests==2.32.3
    # via alibabacloud-tea
rich==14.0.0
    # via
    #   -r requirements.in
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.1
    # via fastapi-cli
rsa==4.9
    # via python-jose
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
slowapi==0.1.9
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy[asyncio]==2.0.40
    # via
    #   -r requirements.in
    #   alembic
starlette==0.46.2
    # via fastapi
structlog==23.3.0
    # via -r requirements.in
tenacity==8.5.0
    # via -r requirements.in
tqdm==4.67.1
    # via openai
typer==0.15.2
    # via fastapi-cli
typing-extensions==4.13.2
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   limits
    #   openai
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   typer
tzdata==2025.2
    # via
    #   pandas
    #   tzlocal
tzlocal==5.3.1
    # via apscheduler
urllib3==2.4.0
    # via requests
uvicorn[standard]==0.29.0
    # via
    #   -r requirements.in
    #   fastapi-cli
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via uvicorn
wrapt==1.17.2
    # via deprecated
yarl==1.19.0
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
